It is now possible to insert the grades of the 5 big Abitur exams and to calculate the grade for the whole Abitur is possible as well now.

+++ app/src/main/java/com/example/schulmanager/models/Fach.java
@@ -55,8 +55,8 @@ public class Fach implements Serializable {
     public void setMuendlich(double punkte) { this.muendlich = punkte; }
 
     // Methode hinzufügen
-    public double getDurchschnittsPunkte() {
-        return (schriftlich + muendlich) / 2;
+    public int getDurchschnittsPunkte() {
+        return (int) ((schriftlich + muendlich) / 2);
     }
     // toString() aktualisieren
     @Override
